msm: kgsl: set vm_pgoff of vma to zero
kgsl gets the entry id or the gpu address through vm_pgoff. It is used
during mmap and never needed again. But this pgoff has different meaning
at other parts of kernel. Not setting to zero will let way for wrong
assumption when tried to unmap a page from the vma.

Change-Id: Ia81c64a77456caf168c6bd23bdf5755c3f3ee31c