msm: kgsl: Defer memory frees done from event kthread
Event work can result in freeing memory as part of free on timestamp.
As freeing memory can take time it shouldn't be done as part of RT
thread as it can result in hogging CPU for longer duration and not
giving chance to other tasks to run. Hence defer memory frees done
from event kthread.

Change-Id: I057ce391030efe9a5795c5cffe6364c6ff732d71
[PA: Ported to 5.10, replaced sched_set_fifo with sched_setscheduler_nocheck]

Change-Id: I09cdac4c79cc5717b0028d6ca464bd509a7b3718